Organized the Dashboard page code by splitting it into different components
