+++ client/src/pages/auth/dashboard/modern/Dashboard.jsx
@@ -2,6 +2,9 @@ import { useEffect, useState } from "react";
 
 // Components
 import SideMenu from "./components/SideMenu";
+import DashboardReports from "./components/DashboardReports";
+import ProjectsTable from "./components/ProjectsTable";
+import TasksTable from "./components/TasksTable";
 
 // Utils
 import fetchUserProjectsUtil from "./utils/fetchUserProjectsUtil";
@@ -19,6 +22,10 @@ import createNewProjectUtil from "./utils/createNewProjectUtil";
 import createNewTaskUtil from "./utils/createNewTaskUtil";
 import updateTaskUtil from "./utils/updateTaskUtil";
 import deleteTaskUtil from "./utils/deleteTaskUtil";
+import ProjectPopup from "./components/ProjectPopup";
+import TaskPopup from "./components/TaskPopup";
+import NewProjectForm from "./components/NewProjectForm";
+import NewTaskForm from "./components/NewTaskForm";
 
 const projectStates = {
   labels: {
@@ -622,132 +629,7 @@ export default function DashboardPageModern({
           <div>
             <section>
               <div className="top-section">
-                <div className="dashboard-reports">
-                  <div className="section projects">
-                    <div className="section-title poppins-bold">Projects</div>
-                    <div className="section-reports">
-                      <div
-                        className={
-                          "dashboard-report" +
-                          (reportsStats.projects.closeToDeadline > 0
-                            ? " close-to-deadline attention"
-                            : "")
-                        }
-                      >
-                        <div className="report-container">
-                          <div className="report-title poppins-medium">
-                            Close to deadline
-                          </div>
-                          <div className="value poppins-bold">
-                            {reportsStats.projects.closeToDeadline}
-                          </div>
-                        </div>
-                      </div>
-                      <div
-                        className={
-                          "dashboard-report" +
-                          (reportsStats.projects.overdue > 0
-                            ? " overdue attention"
-                            : "")
-                        }
-                      >
-                        <div className="report-container">
-                          <div className="report-title poppins-medium">
-                            Overdue
-                          </div>
-                          <div className="value poppins-bold">
-                            {reportsStats.projects.overdue}
-                          </div>
-                        </div>
-                      </div>
-                      <div className="dashboard-report">
-                        <div className="report-container">
-                          <div className="report-title poppins-medium">
-                            Not started
-                          </div>
-                          <div className="value poppins-bold">
-                            {reportsStats.projects.notStarted}
-                          </div>
-                        </div>
-                      </div>
-                      <div className="dashboard-report">
-                        <div className="report-container">
-                          <div className="report-title poppins-medium">
-                            In progress
-                          </div>
-                          <div className="value poppins-bold">
-                            {reportsStats.projects.inProgress}
-                          </div>
-                        </div>
-                      </div>
-                      <div className="dashboard-report completed">
-                        <div className="report-container">
-                          <div className="report-title poppins-medium">
-                            Completed
-                          </div>
-                          <div className="value poppins-bold">
-                            {reportsStats.projects.completed}
-                          </div>
-                        </div>
-                      </div>
-                    </div>
-                  </div>
-                  <div className="section tasks">
-                    <div className="section-title poppins-bold">Tasks</div>
-                    <div className="section-reports">
-                      <div className="dashboard-report today">
-                        <div className="report-container">
-                          <div className="report-title poppins-medium">
-                            Today
-                          </div>
-                          <div className="value poppins-bold">
-                            {reportsStats.tasks.today}
-                          </div>
-                        </div>
-                      </div>
-                      <div className="dashboard-report">
-                        <div className="report-container">
-                          <div className="report-title poppins-medium">
-                            Inbox
-                          </div>
-                          <div className="value poppins-bold">
-                            {reportsStats.tasks.inbox}
-                          </div>
-                        </div>
-                      </div>
-                      <div className="dashboard-report">
-                        <div className="report-container">
-                          <div className="report-title poppins-medium">
-                            To do
-                          </div>
-                          <div className="value poppins-bold">
-                            {reportsStats.tasks.toDo}
-                          </div>
-                        </div>
-                      </div>
-                      <div className="dashboard-report">
-                        <div className="report-container">
-                          <div className="report-title poppins-medium">
-                            Doing
-                          </div>
-                          <div className="value poppins-bold">
-                            {reportsStats.tasks.doing}
-                          </div>
-                        </div>
-                      </div>
-                      <div className="dashboard-report done">
-                        <div className="report-container">
-                          <div className="report-title poppins-medium">
-                            Done
-                          </div>
-                          <div className="value poppins-bold">
-                            {reportsStats.tasks.done}
-                          </div>
-                        </div>
-                      </div>
-                    </div>
-                  </div>
-                </div>
+                <DashboardReports reportsStats={reportsStats} />
               </div>
               <div className="bottom-section">
                 <div className="controls poppins-medium">
@@ -771,120 +653,21 @@ export default function DashboardPageModern({
                   </div>
                 </div>
                 {selectedTable === "tasks" ? (
-                  <table className="tasks-table">
-                    <thead>
-                      <tr>
-                        <th className="poppins-medium">Name</th>
-                        <th className="poppins-medium">State</th>
-                        <th className="poppins-medium">Priority</th>
-                        <th className="poppins-medium">Assignee</th>
-                        <th className="poppins-medium">Project</th>
-                        <th className="poppins-medium">Updated</th>
-                        <th className="poppins-medium">Created</th>
-                      </tr>
-                    </thead>
-                    <tbody>
-                      {userTasks.map((userTask, index) => {
-                        const updated = new Date(
-                          userTask.updated_on,
-                        ).toLocaleString();
-                        const created = new Date(
-                          userTask.created_on,
-                        ).toLocaleString();
-                        const project = userProjects.filter((project) => {
-                          if (project.project_id === userTask.project) {
-                            return project;
-                          }
-                        });
-                        return (
-                          <tr
-                            key={index}
-                            onClick={() => openTaskPopup(userTask.task_id)}
-                          >
-                            <td>{userTask.name}</td>
-                            <td>
-                              <span
-                                className={
-                                  "poppins-semibold " +
-                                  taskStates.classes[userTask.state]
-                                }
-                              >
-                                {taskStates.labels[userTask.state]}
-                              </span>
-                            </td>
-                            <td>{taskPriorities.labels[userTask.priority]}</td>
-                            <td>
-                              {userTask.assigned_to === userId ? "Me" : ""}
-                            </td>
-                            <td>{project.length > 0 ? project[0].name : ""}</td>
-                            <td>{updated}</td>
-                            <td>{created}</td>
-                          </tr>
-                        );
-                      })}
-                    </tbody>
-                  </table>
+                  <TasksTable
+                    userTasks={userTasks}
+                    userProjects={userProjects}
+                    openTaskPopup={openTaskPopup}
+                    taskStates={taskStates}
+                    taskPriorities={taskPriorities}
+                    userId={userId}
+                  />
                 ) : (
-                  <table className="projects-table">
-                    <thead>
-                      <tr>
-                        <th className="poppins-medium">Name</th>
-                        <th className="poppins-medium">State</th>
-                        <th className="poppins-medium">Owner</th>
-                        <th className="poppins-medium">Deadline</th>
-                        <th className="poppins-medium">Updated</th>
-                        <th className="poppins-medium">Created</th>
-                      </tr>
-                    </thead>
-                    <tbody>
-                      {userProjects.map((userProject, index) => {
-                        const deadline = new Date(
-                          userProject.deadline,
-                        ).toLocaleString();
-                        const updated = new Date(
-                          userProject.updated_on,
-                        ).toLocaleString();
-                        const created = new Date(
-                          userProject.created_on,
-                        ).toLocaleString();
-                        return (
-                          <tr
-                            key={index}
-                            onClick={() =>
-                              openProjectPopup(userProject.project_id)
-                            }
-                          >
-                            <td>
-                              <div>
-                                <span
-                                  className={
-                                    projectStates.classes[userProject.state]
-                                  }
-                                >
-                                  <MdOutlineRadioButtonChecked />
-                                </span>
-                                {userProject.name}
-                              </div>
-                            </td>
-                            <td>
-                              <span
-                                className={
-                                  "poppins-semibold " +
-                                  projectStates.classes[userProject.state]
-                                }
-                              >
-                                {projectStates.labels[userProject.state]}
-                              </span>
-                            </td>
-                            <td>{userProject.owner === userId ? "Me" : ""}</td>
-                            <td>{deadline}</td>
-                            <td>{updated}</td>
-                            <td>{created}</td>
-                          </tr>
-                        );
-                      })}
-                    </tbody>
-                  </table>
+                  <ProjectsTable
+                    userProjects={userProjects}
+                    projectStates={projectStates}
+                    openProjectPopup={openProjectPopup}
+                    userId={userId}
+                  />
                 )}
               </div>
               <div className="quick-actions">
@@ -910,506 +693,50 @@ export default function DashboardPageModern({
               </div>
             </section>
             {showProjectPopup ? (
-              <div
-                className={
-                  "project-popup" + (projectPopupVisible ? " visible" : "")
-                }
-              >
-                <div className="project-header">
-                  <div
-                    className={
-                      "project-state poppins-semibold " +
-                      projectStates.classes[popupProject.state]
-                    }
-                  >
-                    {projectStates.labels[popupProject.state]}
-                  </div>
-                  <h3 className="project-name poppins-semibold">
-                    {popupProject.name}
-                  </h3>
-                </div>
-                <div className="project-info">
-                  <div className="owner">
-                    Owned by {popupProject.owner === userId ? "Me" : ""}
-                  </div>
-                  <div className="updated">
-                    Updated{" "}
-                    {updatedMessageUtil(new Date(popupProject.updated_on))}
-                  </div>
-                </div>
-                <div className="project-details">
-                  <div className="description-label poppins-medium">
-                    Description
-                  </div>
-                  <textarea
-                    className="description poppins-regular"
-                    value={projectUpdates.description}
-                    onChange={(e) => updateProjectDescription(e.target.value)}
-                  />
-                </div>
-                <div className="project-tasks">
-                  <h3 className="title">Tasks</h3>
-                  <table className="project-tasks-table">
-                    <thead>
-                      <tr>
-                        <th className="poppins-medium">Name</th>
-                        <th className="poppins-medium">State</th>
-                        <th className="poppins-medium">Assignee</th>
-                        <th className="poppins-medium">Updated</th>
-                        <th className="poppins-medium">Created</th>
-                      </tr>
-                    </thead>
-                    <tbody>
-                      {userTasks.map((task, index) => {
-                        if (task.project === popupProject.project_id) {
-                          const updated = new Date(
-                            task.updated_on,
-                          ).toLocaleString();
-                          const created = new Date(
-                            task.created_on,
-                          ).toLocaleString();
-                          return (
-                            <tr className="project-task" key={index}>
-                              <td>
-                                <div>
-                                  <span
-                                    className={taskStates.classes[task.state]}
-                                  >
-                                    <MdOutlineRadioButtonChecked />
-                                  </span>
-                                  {task.name}
-                                </div>
-                              </td>
-                              <td>
-                                <span
-                                  className={
-                                    "poppins-semibold " +
-                                    taskStates.classes[task.state]
-                                  }
-                                >
-                                  {taskStates.labels[task.state]}
-                                </span>
-                              </td>
-                              <td>{task.assigned_to === userId ? "Me" : ""}</td>
-                              <td>{updated}</td>
-                              <td>{created}</td>
-                            </tr>
-                          );
-                        }
-                      })}
-                    </tbody>
-                  </table>
-                </div>
-                <div className="actions">
-                  <div className="project-actions">
-                    {popupProject.state === 1 ? (
-                      <button
-                        className="action poppins-regular"
-                        onClick={() => updateProjectState(2)}
-                      >
-                        Start
-                      </button>
-                    ) : (
-                      ""
-                    )}
-                    {popupProject.state !== 1 ? (
-                      <button
-                        className="action poppins-regular"
-                        onClick={() => updateProjectState(1)}
-                      >
-                        Reset
-                      </button>
-                    ) : (
-                      ""
-                    )}
-                    {popupProject.state !== 3 ? (
-                      <button
-                        className="action poppins-regular"
-                        onClick={() => updateProjectState(3)}
-                      >
-                        Complete
-                      </button>
-                    ) : (
-                      ""
-                    )}
-                  </div>
-                  <div>
-                    <button
-                      className="action poppins-regular delete"
-                      onClick={() => deleteProjectFn()}
-                    >
-                      Delete
-                    </button>
-                    <button
-                      className={
-                        "action poppins-regular save" +
-                        (Object.keys(projectUpdates).length !== 0
-                          ? " enable-save"
-                          : "")
-                      }
-                      onClick={() => saveUpdates()}
-                    >
-                      Save
-                    </button>
-                  </div>
-                </div>
-                <button
-                  className="close-popup poppins-semibold"
-                  onClick={() => closeProjectPopup()}
-                >
-                  Close
-                </button>
-              </div>
+              <ProjectPopup
+                projectPopupVisible={projectPopupVisible}
+                projectStates={projectStates}
+                popupProject={popupProject}
+                userId={userId}
+                projectUpdates={projectUpdates}
+                updateProjectDescription={updateProjectDescription}
+                userTasks={userTasks}
+                taskStates={taskStates}
+                updateProjectState={updateProjectState}
+                deleteProjectFn={deleteProjectFn}
+                saveUpdates={saveUpdates}
+                closeProjectPopup={closeProjectPopup}
+              />
             ) : showTaskPopup ? (
-              <div
-                className={"task-popup" + (taskPopupVisible ? " visible" : "")}
-              >
-                <div className="task-header">
-                  <div
-                    className={
-                      "task-state poppins-semibold " +
-                      taskStates.classes[popupTask.state]
-                    }
-                  >
-                    {taskStates.labels[popupTask.state]}
-                  </div>
-                  <div className="task-name poppins-semibold">
-                    {popupTask.name}
-                  </div>
-                </div>
-                <div className="task-info poppins-regular">
-                  <div className="assigned-to">
-                    Assigned to {popupTask.assigned_to === userId ? "Me" : ""}
-                  </div>
-                  <div className="updated">
-                    Updated {updatedMessageUtil(new Date(popupTask.updated_on))}
-                  </div>
-                </div>
-                <div className="task-details">
-                  <div className="short-description-section">
-                    <label
-                      htmlFor="short-description"
-                      className="short-description-label poppins-medium"
-                    >
-                      Short description
-                    </label>
-                    <input
-                      type="text"
-                      name="short-description"
-                      className="short-description poppins-regular"
-                      value={popupTask.short_description}
-                      onChange={() => console.log("Test")}
-                    />
-                  </div>
-                  <div className="description-section">
-                    <label
-                      htmlFor="description"
-                      className="description-label poppins-medium"
-                    >
-                      Description
-                    </label>
-                    <textarea
-                      name="description"
-                      className="description poppins-regular"
-                      value={popupTask.description}
-                      onChange={() => console.log("Test")}
-                    />
-                  </div>
-                </div>
-                <div className="actions">
-                  <div className="task-actions">
-                    {popupTask.state === 1 ? (
-                      <button
-                        className="action start poppins-regular"
-                        onClick={() => updateTaskState(2)}
-                      >
-                        Start
-                      </button>
-                    ) : (
-                      ""
-                    )}
-                    {popupTask.state !== 1 ? (
-                      <button
-                        className="action reset poppins-regular"
-                        onClick={() => updateTaskState(1)}
-                      >
-                        Reset
-                      </button>
-                    ) : (
-                      ""
-                    )}
-                    {popupTask.state !== 3 ? (
-                      <button
-                        className="action finish poppins-regular"
-                        onClick={() => updateTaskState(3)}
-                      >
-                        Finish
-                      </button>
-                    ) : (
-                      ""
-                    )}
-                  </div>
-                  {popupTask.project ? (
-                    <div className="project poppins-medium">
-                      Project:{" "}
-                      <span className="parent-project">
-                        {
-                          userProjects.filter(
-                            (project) =>
-                              project.project_id === popupTask.project,
-                          )[0].name
-                        }
-                      </span>
-                    </div>
-                  ) : (
-                    ""
-                  )}
-                  <div>
-                    <button
-                      className="action delete poppins-regular"
-                      onClick={() => deleteTaskFn()}
-                    >
-                      Delete
-                    </button>
-                    <button className="action save poppins-regular">
-                      Save
-                    </button>
-                  </div>
-                </div>
-                <button
-                  className="close-popup poppins-semibold"
-                  onClick={() => closeTaskPopup()}
-                >
-                  Close
-                </button>
-              </div>
+              <TaskPopup
+                taskPopupVisible={taskPopupVisible}
+                taskStates={taskStates}
+                popupTask={popupTask}
+                userId={userId}
+                updateTaskState={updateTaskState}
+                userProjects={userProjects}
+                deleteTaskFn={deleteTaskFn}
+                closeTaskPopup={closeTaskPopup}
+              />
             ) : (
               ""
             )}
             {showNewProjectForm ? (
-              <div
-                className={
-                  "new-project-form poppins-regular" +
-                  (newProjectFormVisible ? " visible" : "")
-                }
-              >
-                <button
-                  className="close-form poppins-semibold"
-                  onClick={() => closeNewProjectForm()}
-                >
-                  Close
-                </button>
-                <h2 className="form-title poppins-bold">New Project</h2>
-                <div>
-                  <div className="project-name-section">
-                    <label
-                      htmlFor="project-name"
-                      className="project-name-label poppins-medium"
-                    >
-                      Name
-                    </label>
-                    <input
-                      type="text"
-                      name="project-name"
-                      className="project-name poppins-regular"
-                      onChange={(e) =>
-                        setNewProject({ ...newProject, name: e.target.value })
-                      }
-                    />
-                  </div>
-                  <div className="project-deadline-section">
-                    <label
-                      htmlFor="project-deadline"
-                      className="project-deadline-label poppins-medium"
-                    >
-                      Deadline
-                    </label>
-                    <input
-                      type="date"
-                      name="project-deadline"
-                      className="project-deadline poppins-regular"
-                      onChange={(e) =>
-                        setNewProject({
-                          ...newProject,
-                          deadline: e.target.value,
-                        })
-                      }
-                      required
-                    />
-                  </div>
-                  <div className="project-description-section">
-                    <label
-                      htmlFor="project-description"
-                      className="project-description-label poppins-medium"
-                    >
-                      Description
-                    </label>
-                    <textarea
-                      name="project-description"
-                      className="project-description poppins-regular"
-                      rows="12"
-                      onChange={(e) =>
-                        setNewProject({
-                          ...newProject,
-                          description: e.target.value,
-                        })
-                      }
-                    />
-                  </div>
-                  <button
-                    className="create-project poppins-semibold"
-                    onClick={() => createNewProjectFn()}
-                  >
-                    Create
-                  </button>
-                </div>
-              </div>
+              <NewProjectForm
+                newProjectFormVisible={newProjectFormVisible}
+                closeNewProjectForm={closeNewProjectForm}
+                newProject={newProject}
+                setNewProject={setNewProject}
+                createNewProjectFn={createNewProjectFn}
+              />
             ) : showNewTaskForm ? (
-              <div
-                className={
-                  "new-task-form poppins-regular" +
-                  (newTaskFromVisible ? " visible" : "")
-                }
-              >
-                <button
-                  className="close-form poppins-semibold"
-                  onClick={() => closeNewTaskForm()}
-                >
-                  Close
-                </button>
-                <h2 className="form-title poppins-bold">New Task</h2>
-                <div>
-                  <div className="task-name-section">
-                    <label
-                      htmlFor="task-name"
-                      className="task-name-label poppins-medium"
-                    >
-                      Name
-                    </label>
-                    <input
-                      type="text"
-                      name="task-name"
-                      className="task-name poppins-regular"
-                      onChange={(e) =>
-                        setNewTask({ ...newTask, name: e.target.value })
-                      }
-                    />
-                  </div>
-                  <div className="task-priority-section">
-                    <label
-                      htmlFor="task-priority"
-                      className="task-priority-label poppins-medium"
-                    >
-                      Priority
-                    </label>
-                    <div>
-                      <div className="high-priority">
-                        <input
-                          type="radio"
-                          name="task-priority"
-                          className="task-priority poppins-regular"
-                          value="1"
-                          onChange={(e) =>
-                            setNewTask({
-                              ...newTask,
-                              priority: Number(e.target.value),
-                            })
-                          }
-                        />
-                        <label
-                          htmlFor="task-priority"
-                          className="high-priority-label poppins-regular"
-                        >
-                          High
-                        </label>
-                      </div>
-                      <div className="medium-priority">
-                        <input
-                          type="radio"
-                          name="task-priority"
-                          className="task-priority poppins-regular"
-                          value="2"
-                          defaultChecked
-                          onChange={(e) =>
-                            setNewTask({
-                              ...newTask,
-                              priority: Number(e.target.value),
-                            })
-                          }
-                        />
-                        <label
-                          htmlFor="task-priority"
-                          className="medium-priority-label poppins-regular"
-                        >
-                          Medium
-                        </label>
-                      </div>
-                      <div className="low-priority">
-                        <input
-                          type="radio"
-                          name="task-priority"
-                          className="task-priority poppins-regular"
-                          value="3"
-                          onChange={(e) =>
-                            setNewTask({
-                              ...newTask,
-                              priority: Number(e.target.value),
-                            })
-                          }
-                        />
-                        <label
-                          htmlFor="task-priority"
-                          className="low-priority-label poppins-regular"
-                        >
-                          Low
-                        </label>
-                      </div>
-                    </div>
-                  </div>
-                  <div className="task-short-description-section">
-                    <label
-                      htmlFor="task-short-description"
-                      className="task-short-description-label poppins-medium"
-                    >
-                      Short description
-                    </label>
-                    <input
-                      type="text"
-                      name="task-short-description"
-                      className="task-short-description poppins-regular"
-                      onChange={(e) =>
-                        setNewTask({
-                          ...newTask,
-                          short_description: e.target.value,
-                        })
-                      }
-                    />
-                  </div>
-                  <div className="task-description-section">
-                    <label
-                      htmlFor="task-description"
-                      className="task-description-label poppins-medium"
-                    >
-                      Description
-                    </label>
-                    <textarea
-                      type="text"
-                      name="task-description"
-                      className="task-description poppins-regular"
-                      rows="12"
-                      onChange={(e) =>
-                        setNewTask({ ...newTask, description: e.target.value })
-                      }
-                    />
-                  </div>
-                  <button
-                    className="create-task poppins-semibold"
-                    onClick={() => createNewTaskFn()}
-                  >
-                    Create
-                  </button>
-                </div>
-              </div>
+              <NewTaskForm
+                newTaskFromVisible={newTaskFromVisible}
+                closeNewTaskForm={closeNewTaskForm}
+                newTask={newTask}
+                setNewTask={setNewTask}
+                createNewTaskFn={createNewTaskFn}
+              />
             ) : (
               ""
             )}
