+++ client/src/pages/auth/dashboard/modern/components/DashboardReport.jsx
@@ -1,0 +1,25 @@
+export default function DashboardReport({
+  title,
+  value,
+  classes,
+  conditionalClasses,
+}) {
+  return (
+    <div
+      className={
+        "dashboard-report" +
+        (classes ? classes : "") +
+        (conditionalClasses
+          ? conditionalClasses.condition
+            ? conditionalClasses.classes
+            : ""
+          : "")
+      }
+    >
+      <div className="report-container">
+        <div className="report-title poppins-medium">{title}</div>
+        <div className="value poppins-bold">{value}</div>
+      </div>
+    </div>
+  );
+}
